requirements.txt
azure-identity
azure-storage-blob
pandas
sqlalchemy
pyodbc